move moto_server dep to compare parquet tests
